#2 Simplify / harden CAS browser authentication flow
The previous usage of filters was faulty in regards of the serialization/
deserialization of user attributes. A crafted request could be used in
order to circumvent the authentication mechanism. The filters were
listening to "/cas/validate". This URL is no longer needed as well.

Instead the whole authentication mechanism points to a new sonar url
which is now handled by the CasIdentifyProvider.java, listening to
`/sessions/init/cas`. The Provider validates directly the the given
ticket (without ser./deres.). The writing to the servlet response is
kept as it was.