#2 Add Support for Single Sign Out
1. On login `CasIdentifyProvider` is the Granting Ticket is stored with
an mapping CAS Granting Ticket to valid JWT token.
1. Each (permitted) request hits `ForceCasLoginFilter` where a JWT black
list is checked if the JWT is still valid.
1. On external logout, CAS notifies the configured URL (which hits
`CasIdentifyProvider`) with a logout info along the original Granting
Ticket ID. This ID is used to invalidate the JWT token and update it in
the store.
1. Each (now prohibited) request hits `ForceCasLoginFilter` still with
a JWT which but now invalid. The request is redirected to a login page