#2 Enable SonarQube-local logout menu entry by calling CAS logout page
Backchannel logout is already supported. Up till now there was no CAS
support for local logout. The logout button would just re-login and the
session would still be valid.

local logout is now achieved by injecting JavaScript into the
frontend with help of a ServletFilter. The frontend is a single page
application so JavaScript must be active in the first place. Said
JavaScript adds a click handler in the logout menu entry. When the user
clicks the menu entry, the click handler redirects the whole browser
page to the CAS logout page. This triggers the usual backchannel logout
process which invalidates the user's session.