#2 Implements independent property fetching mechanism
up til now `AbstractCasFilter` was THE over-engineered go-to for
CAS properties. It relied on an enum `SonarCasProperties`.

Since there is already a class (or specifically an enum in this case)
and I need to take access CAS properties outside of Filters, I mandated
`SonarCasProperties` to be the API to handle ALL CAS properties. That
includes loading the properties file, returning actual values, and
logging an error if the property was not defined.