#2 restore originally requested URL after successful CAS login
Previously, when the requested SonarQube URL was disregarded after a
successful CAS login and led always to the SonarQube home.

This commit introduces a solution to this. Sadly, CAS does not return
a possibly given original URL inside the service URL without further
work. This led to the solution by saving the originally requested URL
inside a cookie which is going to be evalutated and deleted after a
successful login.

Since now URLs different than `/` are acceptable, it is necessary to
inject the Logout javascript in a more appropriate fashion, the problem
being that it was unclear where to inject javascript and where not.

This commit introduces a check if the browser accepts HTML-ish resources.
If so then javascript is going to be injected. A whitelist or resources
would not suffice because customized URLs by other plugins may provide
URLs that are impossible to know at build time.

Since there happen to be a lot of different cookie actions a new utility
class CookieUtil was created which handles all the cookie related
actions.

The maximum age of said cookie is configurable by a new property
`sonar.cas.urlAfterCasRedirectCookieMaxAgeSeconds`.