#2 persisting session tokens into a file base store
Previously there was only an in-memory database for the session storage
which has several drawbacks, f.i. when the SonarQube container is
restarted all the blacklisted tokens are lost

The storage path is configurable by the key `sonar.cas.sessionStorePath`

The property enum was adjusted as appropriate. Also a dead exception
was removed.