#2 Removes proprietary properties file handling in favor of Sonar's
Previously there SonarCasProperties was in charge of fetching the
`sonar.properties` file and interpret it. This was a mistake to begin
with as it jeopardizes sonar property handling in general.

Instead Sonar's `Configuration` class is used to fetch all found
properties. `SonarCasProperties` does now accept a Configuration class
to continue with the defined keys. Two different behaviours exist
therein:

- mustGet{ValueType}(config)
  - methods starting with "must" either return successfully a value of
  the desired type or throw an Exception.
  - this is usually the case when it comes to mandatory properties
- getString(config, defaultValue)
  - this method may return the given default value when the appropriate
  property was not found