#2 Implement JWT token refresh handling; simplifies file handling
JWT tokens can be refreshed, that is: The same JWT keeps all internal
data but the expiry date is slightly longer.

This commit introduces the refresh handling which holds true for the
following circumstance:

1. both request AND response contain a JWT token
2. both JWT tokens have the same ID

The GrantingTicketFileHandler does no longer store a SimpleJwt. Instead
it saves only the JWT ID. This way there is no longer a replace()
method needed.

For this, the JwtProcessor complements with similar methods for JWT
filtering both with throwing an exception and returning a null JWT
object for easy case handling.

All this is started within the `CasTokenRefreshFilter`.