#2 Fixes problem with continued filtering after back-channel logout
Previously it seemed sufficient to remove the cookie if an invalid
JWT token was found. In conjunction with the Cloudogu EcoSystem this
does not suffice.

Now not only the JWT cookie is removed but also a redirect to the
log-in page is conducted

The `ForceCasLoginFilter` reacts to the LogoutHandler result and
continues to filter or aborts the filtering.