#10 Add SessionStore cleanup feature; adds test coverage
This commit introduces a clean-up feature which conducts some house-
keeping for created files.

The FileSessionStore maintains already an expiration date within the
JWT files. In a background task, this expiration date is used to
determine if a set of files (JWT and associated Service Ticket) should
be deleted.

For the removed tests:

Some tests did add close to none substantial value. Getter are not
supposed to be tested.

In a another instance, one unit test from SimpleJwtTest was removed
which may've lead to flaky results, with the problem being to be run in
the very same second. This timing test is unnecessary as already both
major equivalence classes (past, future) are covered.