#11 properly extracts LoginHandler and LogoutHandler, moves constant
Extraction:

The CasIdentityProvider does basically two things:

1. login
2. logout

The functionality of both were spread all over the class while a single
part moved to LogoutHandler a long time ago.

This commit introduces a proper split between both functionalities and
extracts both parts in a proper way.

Furthermore, the constant `JWT_SESSION_COOKIE` moves to the CookieUtil
class. Its usage was spread all over the codebase allthewhile
LogoutHandler was not a proper place for accommodation. CookieUtil is in
this terms a better place.