#13 Decouple Javascript event registering for better local logout
The injected Javascript which logs out the user locally from SonarQube
stopped the event registering once the menu was clicked. Since the
logout link is removed from the DOM when the user closes the menu, a
proper logout is no longer possible.

This commit decouples this behaviour so that now two event handlers are
registered:

1. A click handler for the menu itself
2. A click handler for the logout link

Every time the menu is opened, the logout click handler will be
newly registered so that this handler works for each time, the menu is
opened.