#10 Refactor connection and exception handling
With more recent Java versions, Try-With-Resource is a more concise way
to handle resource handling. This commit introduces this mechanism and
rewrites usages of unstable Google libraries.

Furthermore, the exception handling harmonized so that at top-level
(at worst) only CasAuthenticationExceptions are thrown.

Furthermore, this commit introduces a more consistent style for
file/path object creation. In conjunction with the exception handling
for JAXB runtime exceptions this led to the JwtFileUtil class which now
handles marshalling and unmarshalling of JWT files in a comparable way.