#10 Unite file handler for Session Store manipulation
For one, the GrantingTicketFileHandler did not handle Granting Tickets
but Service Tickets which are issued after validation the Granting
Ticket.

Secondly, the file handling does no longer seem to be a separated
concern for two different classes but rather of a single class. Thus,
GrantingTicketFileHandler and JwtTokenFileHandler are combined into one
single file.

This new class `SessionFileHandler` takes now care of all the file
handling which reflect in the changed method names as well.