#15 Handle authentication only on the server sides
There are 3 internal SonarQube components in which a Plugin can act.
Authentication must only take place on the Server side because we
heavily rely on HTTP request and response communication which is moot
outside that component.

This commit binds sonar extensions only if we are on the server side of
SonarQube.