#13 Stop event bubbling of same logout link
With the injected javascript there are now two click handler on the same
logout link. This creates a rather unwanted race condition.

While we want to show the CAS logout screen it is possible that instead
the original Sonar logout handler loses and is called second, leading to
an unauthenticated Sonar call which chain-reacts into a redirect to the
log-in page instead.

This commit stops the event bubbling of event handlers with the same
element (which is different from event bubbling within stacked elements).
If the injected handler wins, SonarQube's handler is stopped. The other
way round does not matter because calling the CAS logout URL is what we
actually want.