#19 Allow default max cookie age
The maximum cookie age during the redirection from CAS back to Sonar
is not so important to fail when the respective property was not
configured. Instead a default value of 5 minutes will be given which
should suffice even for sluggish systems.

Either way, the cookie content is NOT vital for the Sonar-CAS
interaction. So even if the cookie DID expire, it does not affect the
authentication in any way. Instead, the original dialed URL may lead to
an error page with a link back to the Sonar homepage.