#22 Cache javascript injection-let
Previously every HTML request was modified with a JS file from
the file system. With rising access numbers this will lead to
a higher CPU load.

This commit reads the file once and caches the javascript. This
introduces the evil of cache invalidation which is currently
solved by container restart.