#34 Default to CAS group replication
This commit inverts the previously introduced group replication logic and
defaults to "CAS" group replication mode because there are more users that want
to use CAS's groups than users who want to use SonarQube internal groups.

This feature is considered a breaking change for either CAS or other group
replication modes. The change to the CHANGELOG.md reflects this insight.