#38 Redirect after login to Cookie URL with correct schema
Any unauthenticated SonarQube request will be redirected to the CAS server. The
original URL including query parameters will be temporally preserved in a cookie
in order to save the full URL through the login process (otherwise the full URL
would be lost because of how CAS works)

The login request might contain an unsafe URL if SonarQube (and thus Sonar-CAS-
Plugin) runs inside a network with an SSL-terminating reversed-proxied. The
outside URL would then use the `https://` schema while SonarQube internally only
the `http://` schema would be used.

This commit replaces the schema inside the redirection URL cookie with the URL
that is configured in the `sonar.properties` file to avoid the unsafe redirect.

This also means that the Sonar-Properties key `sonar.cas.sonarServerUrl` must
start with a schema, otherwise an configuration exception will show during
runtime.